% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sineCreator}
\alias{sineCreator}
\title{A Sine Function}
\usage{
sineCreator(min = 1, max = 10, stepsize = 0.1)
}
\arguments{
\item{min}{smallest value in generating the sine sequence (1 is a good start)}

\item{max}{largest value in generating the sine sequence (10 is a good start)}

\item{stepsize}{value that increases between the min and max (.1 is a good start)}
}
\value{
plot of your sine function with x axis "time" and y axis "sine wave"
sinecreator()
}
\description{
This function allows you to make and plot simple Sine waves using the R sin() function.
}
\examples{
sineCreator(1,10,.1) generates a typical sine function you may imagine
cat_function()
}
\keyword{function}
\keyword{sine}
